Improved performance when display movements pause
the compressor holds items in a queue, so it can compress several frames together. This results in an issue when the main display stops moving, (e.g. if user move mouse to external display), because the most recent frames are stuck in the compression queue. For the user, this appears to be a frozen screen.

this commit solve this issue by dumping the compression queue if there has been no movement after a small time period.